backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.1
python-dotenv==1.0.1
pydantic-settings==2.7.1
httpx==0.28.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
google-auth==2.38.0
python-multipart==0.0.20

# Locked Generative AI and Scraping Ecosystem (to permanently stop pip backtracking)
langchain==1.2.14
langgraph==1.1.4
langchain-google-genai==4.1.1
pdfplumber==0.11.9
pdfminer.six==20251230
python-docx==1.2.0
thefuzz[speedup]==0.22.1
google-api-python-client==2.193.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
beautifulsoup4==4.14.3
langchain-groq==1.1.2
websockets==15.0.1
google-genai==1.55.0
numpy==2.4.4

# Vector database
chromadb>=0.5.0
sentence-transformers>=2.7.0

# Caching (optional — in-process cache used as fallback)
redis[hiredis]>=5.0.0
